Move deploy commands to deploy package
It is cleaner and makes more sense for these commands to be in the one
package
